# openwebui-bridge/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Cybersecurity Agent Tool Bridge",
    "description": "Complete cybersecurity agent ecosystem with 68+ tools across 5 categories for Open WebUI integration",
    "version": "1.0.0",
    "contact": {
      "name": "ATTCK Nexus",
      "url": "https://attck.nexus"
    }
  },
  "servers": [
    {
      "url": "https://offsec-team.attck.nexus",
      "description": "Primary production server"
    },
    {
      "url": "https://offsec-team.attck.community.workers.dev",
      "description": "Cloudflare Worker endpoint"
    },
    {
      "url": "https://tools.attck.nexus",
      "description": "Legacy production server"
    },
    {
      "url": "http://localhost:8001", 
      "description": "Development server"
    }
  ],
  "security": [
    {
      "HTTPBearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ToolRequest": {
        "type": "object",
        "required": ["tool_name", "agent"],
        "properties": {
          "tool_name": {
            "type": "string",
            "description": "Name of the tool to execute",
            "examples": ["test_injection_vulnerabilities", "generate_language_template", "launch_automated_scan"]
          },
          "parameters": {
            "type": "object",
            "additionalProperties": true,
            "description": "Tool parameters"
          },
          "agent": {
            "type": "string",
            "enum": ["rt_dev", "bug_hunter", "burpsuite_operator", "daedelu5", "nexus_kamuy"],
            "description": "Target agent"
          },
          "request_id": {
            "type": "string",
            "description": "Optional request ID for tracking"
          }
        }
      },
      "ToolResponse": {
        "type": "object",
        "required": ["success", "agent", "tool_name", "timestamp"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "result": {
            "description": "Tool execution result"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "agent": {
            "type": "string"
          },
          "tool_name": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "execution_time_ms": {
            "type": "integer",
            "nullable": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/execute": {
      "post": {
        "tags": ["Core"],
        "summary": "Execute Cybersecurity Tool",
        "description": "Execute any cybersecurity tool from the available agent ecosystem. Available agents: rt_dev (code generation), bug_hunter (vulnerability scanning), burpsuite_operator (automated penetration testing), daedelu5 (compliance auditing), nexus_kamuy (workflow orchestration).",
        "operationId": "execute_tool_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ToolRequest"},
              "examples": {
                "vulnerability_scan": {
                  "summary": "Web Vulnerability Scan",
                  "description": "Scan a website for SQL injection and XSS vulnerabilities",
                  "value": {
                    "tool_name": "test_injection_vulnerabilities",
                    "agent": "bug_hunter", 
                    "parameters": {
                      "target_url": "https://example.com",
                      "scan_depth": "medium",
                      "include_blind": false
                    }
                  }
                },
                "framework_detection": {
                  "summary": "Detect Web Framework",
                  "description": "Identify the technology stack of a target website",
                  "value": {
                    "tool_name": "detect_framework",
                    "agent": "bug_hunter",
                    "parameters": {
                      "target_url": "https://example.com"
                    }
                  }
                },
                "code_generation": {
                  "summary": "Generate Code Template", 
                  "description": "Generate a FastAPI code template with tests",
                  "value": {
                    "tool_name": "generate_language_template",
                    "agent": "rt_dev",
                    "parameters": {
                      "language": "python",
                      "template_type": "fastapi",
                      "include_tests": true
                    }
                  }
                },
                "infrastructure_deployment": {
                  "summary": "Deploy Infrastructure",
                  "description": "Deploy a Docker compose stack for development",
                  "value": {
                    "tool_name": "deploy_docker_compose_stack",
                    "agent": "rt_dev",
                    "parameters": {
                      "services": ["web", "database", "redis"],
                      "environment": "development"
                    }
                  }
                },
                "burp_scan": {
                  "summary": "BurpSuite Automated Scan",
                  "description": "Run a comprehensive BurpSuite scan",
                  "value": {
                    "tool_name": "launch_automated_scan", 
                    "agent": "burpsuite_operator",
                    "parameters": {
                      "target_url": "https://example.com",
                      "scan_type": "crawl_and_audit",
                      "timeout": 3600
                    }
                  }
                },
                "compliance_audit": {
                  "summary": "Compliance Audit",
                  "description": "Audit infrastructure for SOC2 compliance",
                  "value": {
                    "tool_name": "audit_infrastructure_compliance",
                    "agent": "daedelu5",
                    "parameters": {
                      "framework": "SOC2",
                      "scope": "infrastructure"
                    }
                  }
                },
                "workflow_orchestration": {
                  "summary": "Multi-Agent Workflow",
                  "description": "Create a coordinated workflow across multiple agents",
                  "value": {
                    "tool_name": "create_multi_agent_workflow",
                    "agent": "nexus_kamuy",
                    "parameters": {
                      "workflow_type": "security_assessment",
                      "agents": ["bug_hunter", "burpsuite_operator"],
                      "target": "https://example.com"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tool executed successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ToolResponse"},
                "examples": {
                  "vulnerability_scan_result": {
                    "summary": "Vulnerability scan completed",
                    "value": {
                      "success": true,
                      "result": {
                        "vulnerabilities_found": 3,
                        "scan_status": "completed",
                        "target_url": "https://example.com",
                        "vulnerabilities": [
                          {
                            "type": "XSS",
                            "severity": "medium",
                            "location": "/search?q="
                          }
                        ]
                      },
                      "error": null,
                      "agent": "bug_hunter",
                      "tool_name": "test_injection_vulnerabilities",
                      "request_id": "req_123",
                      "timestamp": "2025-01-18T12:00:00Z",
                      "execution_time_ms": 2500
                    }
                  },
                  "code_generation_result": {
                    "summary": "Code template generated",
                    "value": {
                      "success": true,
                      "result": {
                        "template_created": true,
                        "language": "python",
                        "framework": "fastapi",
                        "files_generated": ["main.py", "requirements.txt", "tests/test_main.py"],
                        "project_structure": {
                          "main.py": "FastAPI application entry point",
                          "requirements.txt": "Python dependencies",
                          "tests/": "Test directory with unit tests"
                        }
                      },
                      "error": null,
                      "agent": "rt_dev",
                      "tool_name": "generate_language_template",
                      "timestamp": "2025-01-18T12:00:00Z",
                      "execution_time_ms": 1200
                    }
                  }
                }
              }
            }
          },
          "400": {"description": "Invalid tool request or parameters"},
          "401": {"description": "Authentication required - provide Bearer token"},
          "422": {"description": "Validation Error"},
          "500": {"description": "Tool execution failed"}
        },
        "security": [
          {"HTTPBearer": []}
        ]
      }
    }
  },
...
}